fixed Markdown for headings
Using the anchors `<a name=.../>` directly before a `#` heading seems not to be rendered correctly with Github. I added a separating line and the result looks better now.

As a side remark: Github automatically creates anchors from headings, so unless you need the anchors to be exactly how you defined them, you can remove the `<a>` tags and just adjust the links in the table of contents. I can create a PR if you want.