clarified what a Unix timestamp means
As far as I can tell, the way I stated it here is what is meant by "Unix time". Official UTC time involves leap seconds, and system time given by any POSIX-like system (calibrated according to UTC), where the time is given as an integer representing seconds (or milliseconds or nanoseconds) since 1970-01-01 00:00:00.000000000 UTC, gives an integer that is *lower* than the *actual* number of seconds that have elapsed since 1970-01-01 00:00:00.000000000 UTC. This isn't all bad. At any point in time *between actual leap seconds*, if you know the Unix timestamp (as an integer), you can easily figure out what the civil time (as we humans talk about it) is: convert to civil time using the fact that there are 86400 seconds in a day, and leap days occur every four years except for in years divisible by 100 but not by 400.

As for Unix timestamps that are ambiguous (falling on a leap second), it appears that most systems by default give the *greater* instant in time that the Unix timestamp can represent: so 1230768000 seconds means 2009-01-01 00:00:00 UTC, not the earlier second of 2008-12-31 23:59:60 UTC. And the Unix timestamp 1230767999 obviously refers to 2008-12-31 23:59:59 UTC, with no ambiguity.