Make spec versioning explicit
The relationship of the old and new specs was not entirely clear, and the
spec versions did not have names to refer to them by; implementors can now
indicate which version they support, if needed.

This commit establishes the year of release as the version name, makes
it clear that the 2013 version is the current and stable one, and hints
that support for the 2011 version is not necessary in new implementations.

(Also includes some minor language clarifications.)